学习强国/requirements.txt
Pillow==9.4.0
PyQRCode==1.2.1
pytest==6.2.4
pyzbar==0.1.9
selenium==4.7.2
